move examples to a sub-package
